+++ crest/Assets/Crest/Crest/Scripts/LodData/RegisterFlowInput.cs
@@ -12,7 +12,7 @@ namespace Crest
     [ExecuteAlways]
     [AddComponentMenu(MENU_PREFIX + "Flow Input")]
     [HelpURL("https://crest.readthedocs.io/en/stable/user/ocean-simulation.html#flow")]
-    public class RegisterFlowInput : RegisterLodDataInputDisplacementCorrection<LodDataMgrFlow>
+    public class RegisterFlowInput : RegisterLodDataInputWithSplineSupport<LodDataMgrFlow, SplinePointDataFlow>
     {
         public override bool Enabled => true;
 
@@ -21,6 +21,16 @@ namespace Crest
         protected override Color GizmoColor => new Color(0f, 0f, 1f, 0.5f);
 
         protected override string ShaderPrefix => "Crest/Inputs/Flow";
+
+        protected override bool FollowHorizontalMotion => _followHorizontalMotion;
+
+        protected override string SplineShaderName => "Hidden/Crest/Inputs/Flow/Spline Geometry";
+        protected override Vector2 DefaultCustomData => new Vector2(SplinePointDataFlow.k_defaultSpeed, 0f);
+
+        [Header("Other Settings")]
+
+        [SerializeField, Tooltip(k_displacementCorrectionTooltip)]
+        bool _followHorizontalMotion = false;
 
 #if UNITY_EDITOR
         protected override string FeatureToggleName => "_createFlowSim";