+++ crest/Assets/Crest/Crest/Scripts/LodData/RegisterHeightInput.cs
@@ -2,30 +2,21 @@
 
 // This file is subject to the MIT License as seen in the root of this folder structure (LICENSE)
 
+using UnityEditor;
 using UnityEngine;
 
 namespace Crest
 {
     /// <summary>
-    /// Registers a custom input to the wave shape. Attach this GameObjects that you want to render into the displacmeent textures to affect ocean shape.
+    /// Registers a custom input to affect the water height.
     /// </summary>
     [ExecuteAlways]
-    [AddComponentMenu(MENU_PREFIX + "Animated Waves Input")]
-    [HelpURL("https://crest.readthedocs.io/en/stable/user/ocean-simulation.html#animated-waves")]
-    public class RegisterAnimWavesInput : RegisterLodDataInputWithSplineSupport<LodDataMgrAnimWaves>
+    [AddComponentMenu(MENU_PREFIX + "Height Input")]
+    public class RegisterHeightInput : RegisterLodDataInputWithSplineSupport<LodDataMgrAnimWaves>
     {
         public override bool Enabled => true;
 
-        [Header("Anim Waves Input Settings")]
-        [SerializeField, Tooltip("Which octave to render into, for example set this to 2 to use render into the 2m-4m octave. These refer to the same octaves as the wave spectrum editor. Set this value to 0 to render into all LODs.")]
-        float _octaveWavelength = 0f;
-        public override float Wavelength
-        {
-            get
-            {
-                return _octaveWavelength;
-            }
-        }
+        public override float Wavelength => 0f;
 
         public readonly static Color s_gizmoColor = new Color(0f, 1f, 0f, 0.5f);
         protected override Color GizmoColor => s_gizmoColor;
@@ -35,14 +26,11 @@ namespace Crest
         protected override string SplineShaderName => "Crest/Inputs/Animated Waves/Set Base Water Height Using Geometry";
         protected override Vector2 DefaultCustomData => Vector2.zero;
 
-        [SerializeField, Tooltip(k_displacementCorrectionTooltip)]
-        bool _followHorizontalMotion = true;
-        protected override bool FollowHorizontalMotion => _followHorizontalMotion;
+        protected override bool FollowHorizontalMotion => true;
 
+        [Header("Height Input Settings")]
         [SerializeField, Tooltip("Inform ocean how much this input will displace the ocean surface vertically. This is used to set bounding box heights for the ocean tiles.")]
         float _maxDisplacementVertical = 0f;
-        [SerializeField, Tooltip("Inform ocean how much this input will displace the ocean surface horizontally. This is used to set bounding box widths for the ocean tiles.")]
-        float _maxDisplacementHorizontal = 0f;
 
         [SerializeField, Tooltip("Use the bounding box of an attached renderer component to determine the max vertical displacement.")]
         [PredicatedField(typeof(MeshRenderer))]
@@ -60,7 +48,7 @@ namespace Crest
             var maxDispVert = _maxDisplacementVertical;
 
             // let ocean system know how far from the sea level this shape may displace the surface
-            if (_reportRendererBoundsToOceanSystem)
+            if (_reportRendererBoundsToOceanSystem && _renderer != null)
             {
                 var minY = _renderer.bounds.min.y;
                 var maxY = _renderer.bounds.max.y;
@@ -68,10 +56,15 @@ namespace Crest
                 maxDispVert = Mathf.Max(maxDispVert, Mathf.Abs(seaLevel - minY), Mathf.Abs(seaLevel - maxY));
             }
 
-            if (_maxDisplacementHorizontal > 0f || maxDispVert > 0f)
+            if (maxDispVert > 0f)
             {
-                OceanRenderer.Instance.ReportMaxDisplacementFromShape(_maxDisplacementHorizontal, maxDispVert, 0f);
+                OceanRenderer.Instance.ReportMaxDisplacementFromShape(0f, maxDispVert, 0f);
             }
         }
+
+#if UNITY_EDITOR
+        // Animated waves are always enabled
+        protected override bool FeatureEnabled(OceanRenderer ocean) => true;
+#endif // UNITY_EDITOR
     }
 }