+++ crest/Assets/Crest/Crest/Scripts/Spline/SplinePoint.cs
@@ -7,6 +7,11 @@ using UnityEngine;
 
 namespace Crest.Spline
 {
+    public interface ISplinePointCustomData
+    {
+        Vector2 GetData();
+    }
+
     /// <summary>
     /// Spline point, intended to be child of Spline object
     /// </summary>
@@ -33,7 +38,7 @@ namespace Crest.Spline
 
         void OnDrawGizmosSelected()
         {
-            if (transform.parent.TryGetComponent<IReceiveSplinePointOnDrawGizmosSelectedMessages>(out var receiver))
+            if (transform.parent.TryGetComponent(out IReceiveSplinePointOnDrawGizmosSelectedMessages receiver))
             {
                 receiver.OnSplinePointDrawGizmosSelected(this);
             }
@@ -62,6 +67,16 @@ namespace Crest.Spline
             {
                 EditorGUILayout.HelpBox("Spline component must be present on parent of this GameObject.", MessageType.Error);
                 return;
+            }
+
+            // For any components on spline that want custom data added to spline points, add them
+            var customDatas = parent.GetComponents<ISplinePointCustomDataSetup>();
+            foreach (var customData in customDatas)
+            {
+                if (customData.AttachDataToSplinePoint(thisSP.gameObject))
+                {
+                    EditorUtility.SetDirty(thisSP.gameObject);
+                }
             }
 
             GUILayout.Label("Selection", EditorStyles.boldLabel);
@@ -135,18 +150,24 @@ namespace Crest.Spline
             }
         }
 
-        static GameObject CreateNewSP()
+        static GameObject CreateNewSP(Transform spline)
         {
             var newPoint = new GameObject();
             newPoint.name = "SplinePoint";
             newPoint.AddComponent<SplinePoint>();
+            newPoint.transform.parent = spline;
+
+            if (spline.TryGetComponent(out ISplinePointCustomDataSetup customData))
+            {
+                customData.AttachDataToSplinePoint(newPoint);
+            }
+
             return newPoint;
         }
 
         public static GameObject AddSplinePointBefore(Transform parent, int beforeIdx = 0)
         {
-            var newPoint = CreateNewSP();
-            newPoint.transform.parent = parent;
+            var newPoint = CreateNewSP(parent);
 
             // Put in front of child at beforeIdx
             newPoint.transform.SetSiblingIndex(beforeIdx);
@@ -186,8 +207,7 @@ namespace Crest.Spline
             // If no index specified, assume adding after last point
             if (afterIdx == -1) afterIdx = parent.childCount - 1;
 
-            var newPoint = CreateNewSP();
-            newPoint.transform.parent = parent;
+            var newPoint = CreateNewSP(parent);
 
             var newIdx = afterIdx + 1;
             newPoint.transform.SetSiblingIndex(newIdx);